Stat decoupling cluster 4: Defense flat-int -> capped %
Defense converts from flat-int subtraction (dmg -= Blocked) to percentage reduction
(dmg *= (1 - reduction)), capped at UNIVERSAL_STAT_CAP (0.5 = 50%). CalculateFlatDefense /
GetEvolutionModifiedFlatDefense / GetDefenderFlatDefense / FDamageResult.DefenderFlatDefense all
int32->float. Step 5 in DamageCalculator multiplies instead of subtracts; DamageBlockedByDefense now =
HP actually removed.

Restores Defense: cluster-1's re-derived DEFENSE_PER_POINT (~0.0036) produced a 0-0.5 value the old
RoundToInt floored to 0 for nearly every build — Defense was a silent no-op. This both restyles and
revives it. % reduction per impact suits the per-impact damage system.

BonusDefense flat-int gear field deferred (no % meaning — TODO; gear deferred). Functions NOT renamed
to *DefenseReduction (BlueprintPure reflected symbols + LFS .uasset graphs + existing FunctionRedirect
= unverifiable rename risk; semantics changed in place, rename-TODO left for a BP-aware pass). PIE-
verified: low/mid Defense visible reduction, maxed caps 50%.

+++ Source/world_of_refraction/Public/Character/CharacterDataComponent.h
@@ -278,8 +278,11 @@ public:
     UFUNCTION(BlueprintPure, Category = "Combat|Stats")
     float GetEvolutionModifiedCritChance() const;
 
+    // Crystal-aware defense REDUCTION fraction [0, 0.5] (cluster 4: flat-int -> capped %).
+    // TODO: rename to GetEvolutionModifiedDefenseReduction in a Blueprint-aware pass (has an
+    // existing FunctionRedirect — a rename must re-point it).
     UFUNCTION(BlueprintPure, Category = "Combat|Stats")
-    int32 GetEvolutionModifiedFlatDefense() const;
+    float GetEvolutionModifiedFlatDefense() const;
 
     /** Fully-layered SpellDamage scalar: innate (GetEvolutionModifiedSpellDamage)
      *  + equipment BonusSpellDamage, then ×stone ×transient. This is the SAME value
